refactor: rename common helpers
